src/main/java/com/example/producdemo/service/ProductService.java
package com.example.producdemo.service;

import com.example.producdemo.domain.Product;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService {

    @Autowired
    private ProductRepository productRepository;

    public void saveProduct(String name) {
        Product product = new Product();
        product.setName(name);
        productRepository.save(product);
    }

    public List<Product> findByName(String name) {
        return productRepository.findByName(name);
    }


}
